Truncated Exponential Family

TruncExpFam is an R package to handle truncated members from the exponential family.

Details

This package offers truncated counterparts of the density-, distribution-, quantile- and sampling-functions for a broad range of distributions from the exponential family, as implemented in the stats package.

The package also provides functions for estimating the parameters of the distributions from data, given the truncation limits.

For more info, please check rtrunc(), dtrunc() and print.trunc(). Counterparts for density and probability functions are on the roadmap for a future release.

Note

Found a bug? Want to suggest a feature? Contribute to the scientific and open source communities by opening an issue on our home page. Check the "BugReports" field on packageDescription("TruncExpFam") for the URL.

Supported distributions

Beta

Binomial

Chi-Square

Continuous Bernoulli

Exponential

Gamma

Inverse Gamma

Inverse Gaussian

Log-normal

Negative Binomial

Normal

Poisson

See also

Author

Maintainer: Waldir Leoncio w.l.netto@medisin.uio.no

Authors:

René Holst rene.holst@medisin.uio.no