ML Estimation of Distribution Parameters

ML-estimation of the parameters of the distribution of the specified family, truncated at y.min and y.max

Usage

mlEstimationTruncDist(
  y,
  y.min = attr(y, "truncation_limits")$a,
  y.max = attr(y, "truncation_limits")$b,
  tol = 1e-05,
  max.it = 100,
  delta = 0.33,
  print.iter = 0,
  ny = 100,
  ...
)

Arguments

y

Sequence spanning the domain of the truncated distribution

y.min

Lower bound for y

y.max

Upper bound for y

tol

Error tolerance for parameter estimation

max.it

Maximum number of iterations

delta

Indirectly, the difference between consecutive iterations to compare with the error tolerance

print.iter

Determines the frequency of printing (i.e., prints every print.iter iterations)

ny

size of intermediate y range sequence. Higher values yield better estimations but slower iterations

...

other parameters passed to subfunctions

Value

A vector of class trunc_* containing the maximum-likelihood estmation of the underlying distribution * parameters.

Note

print.iter can be TRUE, FALSE or an integer indicating an interval for printing every X iterations.

References

Inspired by Salvador: Pueyo: "Algorithm for the maximum likelihood estimation of the parameters of the truncated normal and lognormal distributions"

Author

René Holst

Examples

sample_size <- 1000
# Normal
sample.norm <- rtrunc(n = sample_size, mean = 2, sd = 1.5, a = -1)
mlEstimationTruncDist(
  sample.norm,
  y.min = -1, max.it = 500, delta = 0.33,
  print.iter = TRUE
)
#> Estimating parameters for the normal distribution
#> it:   1 delta:   0.001349 parm:  2.067 1.385 
#> it:   2 delta:  0.0007173 parm:  2.052 1.404 
#> it:   3 delta:  0.0004003 parm:  2.04 1.418 
#> it:   4 delta:  0.0002313 parm:  2.03 1.428 
#> it:   5 delta:  0.0001371 parm:  2.023 1.437 
#> it:   6 delta:  8.281e-05 parm:  2.017 1.443 
#> it:   7 delta:  5.076e-05 parm:  2.013 1.448 
#> it:   8 delta:  3.147e-05 parm:  2.009 1.452 
#> it:   9 delta:  1.968e-05 parm:  2.007 1.456 
#> it:  10 delta:  1.239e-05 parm:  2.004 1.458 
#> it:  11 delta:  7.844e-06 parm:  2.003 1.46 
#> [1] 2.001301 1.461855

# Log-Normal
sample.lognorm <- rtrunc(
  n = sample_size, family = "lognormal", meanlog = 2.5, sdlog = 0.5, a = 7
)
ml_lognormal <- mlEstimationTruncDist(
  sample.lognorm,
  y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3,
  print.iter = FALSE
)
ml_lognormal
#> [1] 2.4693866 0.4973355

# Poisson
sample.pois <- rtrunc(n = sample_size, lambda = 10, a = 4, family = "Poisson")
mlEstimationTruncDist(
  sample.pois,
  y.min = 4, max.it = 500, delta = 0.33,
  print.iter = 5
)
#> Estimating parameters for the poisson distribution
#> [1] 10.05904

# Gamma
sample.gamma <- rtrunc(n = sample_size, shape = 6, rate = 2, a = 2, family = "Gamma")
mlEstimationTruncDist(
  sample.gamma,
  y.min = 2, max.it = 1500, delta = 0.3,
  print.iter = 10
)
#> Estimating parameters for the gamma distribution
#> it:  10 delta:    0.01745 parm:  7.863 2.457 
#> it:  20 delta:    0.00338 parm:  6.971 2.233 
#> it:  30 delta:  0.0009542 parm:  6.544 2.124 
#> it:  40 delta:  0.0003278 parm:  6.308 2.064 
#> it:  50 delta:  0.0001255 parm:  6.165 2.027 
#> it:  60 delta:  5.122e-05 parm:  6.076 2.004 
#> it:  70 delta:  2.176e-05 parm:  6.019 1.99 
#> it:  80 delta:  9.483e-06 parm:  5.981 1.98 
#> [1] 5.977916 1.979224

# Negative binomial
sample.nbinom <- rtruncnbinom(sample_size, size = 50, prob = .3, a = 100, b = 120)
mlEstimationTruncDist(sample.nbinom, r=10)
#> [1] 110.2314