ML Estimation of Distribution Parameters

ML-estimation of the parameters of the distribution of the specified family, truncated at y.min and y.max

Usage

mlEstimationTruncDist(
  y,
  y.min = attr(y, "truncation_limits")$a,
  y.max = attr(y, "truncation_limits")$b,
  tol = 1e-05,
  max.it = 100,
  delta = 0.33,
  print.iter = 0,
  ny = 100,
  family = NULL,
  ...
)

Arguments

y

Sequence spanning the domain of the truncated distribution

y.min

Lower bound for y

y.max

Upper bound for y

tol

Error tolerance for parameter estimation

max.it

Maximum number of iterations

delta

Indirectly, the difference between consecutive iterations to compare with the error tolerance

print.iter

Determines the frequency of printing (i.e., prints every print.iter iterations)

ny

size of intermediate y range sequence. Higher values yield better estimations but slower iterations

family

distribution family to use

...

other parameters passed to subfunctions

Value

A vector of class trunc_* containing the maximum-likelihood estmation of the underlying distribution * parameters.

Details

If print.iter = TRUE, the function prints the iteration, the sum of squares of delta.eta.j (delta.L2), and the current parameter estimates. The delta argument of this function is a factor in the calculation of delta.eta.j, which in turn is a factor in the calculation of delta.L2.

References

Inspired by Salvador: Pueyo: "Algorithm for the maximum likelihood estimation of the parameters of the truncated normal and lognormal distributions"

Author

René Holst

Examples

sample_size <- 1000
# Normal
sample.norm <- rtrunc(n = sample_size, mean = 2, sd = 1.5, a = -1)
mlEstimationTruncDist(
  sample.norm,
  y.min = -1, max.it = 500, delta = 0.33,
  print.iter = TRUE
)
#> Estimating parameters for the normal distribution
#>  it	 delta.L2	 parameter(s)
#>  1 	  0.001395 	 2.088 1.407 
#>  2 	 0.0007434 	 2.071 1.427 
#>  3 	 0.0004165 	 2.058 1.441 
#>  4 	 0.0002419 	 2.048 1.453 
#>  5 	 0.0001442 	 2.04 1.461 
#>  6 	  8.77e-05 	 2.034 1.468 
#>  7 	 5.414e-05 	 2.029 1.474 
#>  8 	 3.382e-05 	 2.025 1.478 
#>  9 	 2.132e-05 	 2.022 1.482 
#> 10 	 1.354e-05 	 2.02 1.484 
#> 11 	 8.644e-06 	 2.018 1.487 
#>     mean       sd 
#> 2.016520 1.488316 

# Log-Normal
sample.lognorm <- rtrunc(
  n = sample_size, family = "lognormal", meanlog = 2.5, sdlog = 0.5, a = 7
)
ml_lognormal <- mlEstimationTruncDist(
  sample.lognorm,
  y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3,
  print.iter = FALSE
)
ml_lognormal
#>   meanlog     sdlog 
#> 2.5591788 0.4725295 

# Poisson
sample.pois <- rtrunc(
 n = sample_size, lambda = 10, a = 4, family = "Poisson"
)
mlEstimationTruncDist(
  sample.pois,
  y.min = 4, max.it = 500, delta = 0.33,
  print.iter = 5
)
#> Estimating parameters for the poisson distribution
#>  it	 delta.L2	 parameter(s)
#> lambda 
#> 10.056 

# Gamma
sample.gamma <- rtrunc(
 n = sample_size, shape = 6, rate = 2, a = 2, family = "Gamma"
)
mlEstimationTruncDist(
  sample.gamma,
  y.min = 2, max.it = 1500, delta = 0.3,
  print.iter = 10
)
#> Estimating parameters for the gamma distribution
#>  it	 delta.L2	 parameter(s)
#>  10 	   0.01796 	 7.963 2.492 
#>  20 	  0.003439 	 7.06 2.264 
#>  30 	 0.0009639 	 6.63 2.154 
#>  40 	 0.0003287 	 6.393 2.093 
#>  50 	 0.0001249 	 6.251 2.057 
#>  60 	 5.058e-05 	 6.162 2.034 
#>  70 	 2.131e-05 	 6.105 2.019 
#>  80 	 9.207e-06 	 6.067 2.01 
#>    shape     rate 
#> 6.064502 2.008946 

# Negative binomial
sample.nbinom <- rtruncnbinom(
 sample_size, size = 50, prob = .3, a = 100, b = 120
)
mlEstimationTruncDist(sample.nbinom, r=10)
#> Warning: Maximum number of iterations reached. Convergence is not guaranteed. You might want to run again with a higher value for max.it
#>     mean 
#> 65.26543