ML Estimation of Distribution Parameters

ML-estimation of the parameters of the distribution of the specified family, truncated at y.min and y.max

Usage

mlEstimationTruncDist(
  y,
  y.min = attr(y, "truncation_limits")$a,
  y.max = attr(y, "truncation_limits")$b,
  tol = 1e-05,
  max.it = 100,
  delta = 0.33,
  print.iter = 0,
  ny = 100,
  family = NULL,
  ...
)

Arguments

y

Sequence spanning the domain of the truncated distribution

y.min

Lower bound for y

y.max

Upper bound for y

tol

Error tolerance for parameter estimation

max.it

Maximum number of iterations

delta

Indirectly, the difference between consecutive iterations to compare with the error tolerance

print.iter

Determines the frequency of printing (i.e., prints every print.iter iterations)

ny

size of intermediate y range sequence. Higher values yield better estimations but slower iterations

family

distribution family to use

...

other parameters passed to subfunctions

Value

A vector of class trunc_* containing the maximum-likelihood estmation of the underlying distribution * parameters.

Note

print.iter can be TRUE, FALSE or an integer indicating an interval for printing every X iterations.

References

Inspired by Salvador: Pueyo: "Algorithm for the maximum likelihood estimation of the parameters of the truncated normal and lognormal distributions"

Author

René Holst

Examples

sample_size <- 1000
# Normal
sample.norm <- rtrunc(n = sample_size, mean = 2, sd = 1.5, a = -1)
mlEstimationTruncDist(
  sample.norm,
  y.min = -1, max.it = 500, delta = 0.33,
  print.iter = TRUE
)
#> Estimating parameters for the normal distribution
#> it:  1 delta:   0.001317 parm: 2.107 1.401 
#> it:  2 delta:     0.0007 parm: 2.091 1.42 
#> it:  3 delta:  0.0003904 parm: 2.079 1.434 
#> it:  4 delta:  0.0002254 parm: 2.07 1.445 
#> it:  5 delta:  0.0001334 parm: 2.063 1.453 
#> it:  6 delta:  8.053e-05 parm: 2.057 1.46 
#> it:  7 delta:   4.93e-05 parm: 2.052 1.465 
#> it:  8 delta:  3.053e-05 parm: 2.049 1.469 
#> it:  9 delta:  1.907e-05 parm: 2.046 1.472 
#> it: 10 delta:  1.199e-05 parm: 2.044 1.474 
#> it: 11 delta:  7.579e-06 parm: 2.042 1.477 
#>     mean       sd 
#> 2.040792 1.478154 

# Log-Normal
sample.lognorm <- rtrunc(
  n = sample_size, family = "lognormal", meanlog = 2.5, sdlog = 0.5, a = 7
)
ml_lognormal <- mlEstimationTruncDist(
  sample.lognorm,
  y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3,
  print.iter = FALSE
)
ml_lognormal
#>   meanlog     sdlog 
#> 2.4922992 0.4916553 

# Poisson
sample.pois <- rtrunc(
 n = sample_size, lambda = 10, a = 4, family = "Poisson"
)
mlEstimationTruncDist(
  sample.pois,
  y.min = 4, max.it = 500, delta = 0.33,
  print.iter = 5
)
#> Estimating parameters for the poisson distribution
#>   lambda 
#> 9.914884 

# Gamma
sample.gamma <- rtrunc(
 n = sample_size, shape = 6, rate = 2, a = 2, family = "Gamma"
)
mlEstimationTruncDist(
  sample.gamma,
  y.min = 2, max.it = 1500, delta = 0.3,
  print.iter = 10
)
#> Estimating parameters for the gamma distribution
#> it:  10 delta:    0.01848 parm: 8.37 2.621 
#> it:  20 delta:   0.003384 parm: 7.462 2.39 
#> it:  30 delta:  0.0009041 parm: 7.041 2.281 
#> it:  40 delta:  0.0002921 parm: 6.813 2.223 
#> it:  50 delta:  0.0001046 parm: 6.681 2.188 
#> it:  60 delta:  3.972e-05 parm: 6.601 2.167 
#> it:  70 delta:  1.563e-05 parm: 6.551 2.154 
#>    shape     rate 
#> 6.530127 2.149169 

# Negative binomial
sample.nbinom <- rtruncnbinom(
 sample_size, size = 50, prob = .3, a = 100, b = 120
)
mlEstimationTruncDist(sample.nbinom, r=10)
#> Warning: Maximum number of iterations reached. Convergence is not guaranteed. You might want to run again with a higher value for max.it
#>     mean 
#> 62.87075