ML Estimation of Distribution Parameters

ML-estimation of the parameters of the distribution of the specified family, truncated at y.min and y.max

mlEstimationTruncDist(
  y,
  y.min = attr(y, "truncation_limits")$a,
  y.max = attr(y, "truncation_limits")$b,
  tol = 1e-05,
  max.it = 100,
  delta = 0.33,
  print.iter = 0,
  ny = 100,
  ...
)

Arguments

y

Sequence spanning the domain of the truncated distribution

y.min

Lower bound for y

y.max

Upper bound for y

tol

Error tolerance for parameter estimation

max.it

Maximum number of iterations

delta

Indirectly, the difference between consecutive iterations to compare with the error tolerance

print.iter

Determines the frequency of printing (i.e., prints every print.iter iterations)

ny

size of intermediate y range sequence. Higher values yield better estimations but slower iterations

...

other parameters passed to subfunctions

Value

A vector of class trunc_* containing the maximum-likelihood estmation of the underlying distribution * parameters.

Note

print.iter can be TRUE, FALSE or an integer indicating an interval for printing every X iterations.

References

Inspired by Salvador: Pueyo: "Algorithm for the maximum likelihood estimation of the parameters of the truncated normal and lognormal distributions"

Author

René Holst

Examples

sample_size <- 1000
# Normal
sample.norm <- rtrunc(n = sample_size, mean = 2, sd = 1.5, a = -1)
mlEstimationTruncDist(
  sample.norm,
  y.min = -1, max.it = 500, delta = 0.33,
  print.iter = TRUE
)
#> Estimating parameters for the normal distribution
#> it:   1 delta:   0.001599 parm:  2.063 1.437 
#> it:   2 delta:  0.0008586 parm:  2.044 1.459 
#> it:   3 delta:   0.000487 parm:  2.028 1.476 
#> it:   4 delta:  0.0002873 parm:  2.016 1.489 
#> it:   5 delta:  0.0001745 parm:  2.007 1.499 
#> it:   6 delta:  0.0001083 parm:  1.999 1.507 
#> it:   7 delta:  6.839e-05 parm:  1.993 1.514 
#> it:   8 delta:  4.375e-05 parm:  1.988 1.519 
#> it:   9 delta:  2.829e-05 parm:  1.984 1.524 
#> it:  10 delta:  1.844e-05 parm:  1.981 1.527 
#> it:  11 delta:   1.21e-05 parm:  1.979 1.53 
#> it:  12 delta:  7.981e-06 parm:  1.977 1.532 
#> [1] 1.974989 1.533984

# Log-Normal
sample.lognorm <- rtrunc(
  n = sample_size, family = "lognormal", meanlog = 2.5, sdlog = 0.5, a = 7
)
ml_lognormal <- mlEstimationTruncDist(
  sample.lognorm,
  y.min = 7, max.it = 500, tol = 1e-10, delta = 0.3,
  print.iter = FALSE
)
ml_lognormal
#> [1] 2.5479124 0.4755499

# Poisson
sample.pois <- rtrunc(n = sample_size, lambda = 10, a = 4, family = "Poisson")
mlEstimationTruncDist(
  sample.pois,
  y.min = 4, max.it = 500, delta = 0.33,
  print.iter = 5
)
#> Estimating parameters for the poisson distribution
#> [1] 9.911837

# Gamma
sample.gamma <- rtrunc(n = sample_size, shape = 6, rate = 2, a = 2, family = "Gamma")
mlEstimationTruncDist(
  sample.gamma,
  y.min = 2, max.it = 1500, delta = 0.3,
  print.iter = 10
)
#> Estimating parameters for the gamma distribution
#> it:  10 delta:     0.0208 parm:  7.788 2.478 
#> it:  20 delta:   0.004163 parm:  6.81 2.23 
#> it:  30 delta:   0.001252 parm:  6.331 2.106 
#> it:  40 delta:   0.000465 parm:  6.056 2.035 
#> it:  50 delta:  0.0001945 parm:  5.884 1.991 
#> it:  60 delta:  8.749e-05 parm:  5.77 1.961 
#> it:  70 delta:  4.126e-05 parm:  5.694 1.941 
#> it:  80 delta:  2.009e-05 parm:  5.641 1.927 
#> it:  90 delta:  9.996e-06 parm:  5.603 1.918 
#> [1] 5.600286 1.917017

# Negative binomial
sample.nbinom <- rtruncnbinom(sample_size, size = 50, prob = .3, a = 100, b = 120)
mlEstimationTruncDist(sample.nbinom, r=10)
#> [1] 110.4637